+++ api/python/tests/test_admin_api.py
@@ -38,6 +38,10 @@ SSO_CONFIG = {
     "timestamp": datetime.datetime(2024, 6, 14, 11, 42, 27, 857128, tzinfo=datetime.timezone.utc),
     "uploader": USER,
 }
+TABULATOR_TABLE = {
+    "name": "table",
+    "config": "config",
+}
 MUTATION_ERRORS = (
     (
         {
@@ -359,3 +363,63 @@ def test_sso_config_set(data, result):
                 admin.sso_config.set("")
         else:
             assert admin.sso_config.set("") == result
+
+
+@pytest.mark.parametrize(
+    "data, result",
+    [
+        ({"tabulator_tables": [TABULATOR_TABLE]}, [admin.TabulatorTable(**TABULATOR_TABLE)]),
+        (None, admin.BucketNotFoundError),
+    ],
+)
+def test_tabulator_list(data, result):
+    with mock_client(
+        _make_nested_dict("bucket_config", data),
+        "bucketTabulatorTablesList",
+        variables={"name": "test"},
+    ):
+        if isinstance(result, type) and issubclass(result, Exception):
+            with pytest.raises(result):
+                admin.tabulator.list_tables("test")
+        else:
+            assert admin.tabulator.list_tables("test") == result
+
+
+@pytest.mark.parametrize(
+    "data,result",
+    [
+        ({"__typename": "BucketConfig"}, None),
+        *MUTATION_ERRORS,
+    ],
+)
+def test_tabulator_set(data, result):
+    with mock_client(
+        _make_nested_dict("admin.bucket_set_tabulator_table", data),
+        "bucketTabulatorTableSet",
+        variables={"bucketName": "test", "tableName": "table", "config": ""},
+    ):
+        if isinstance(result, type) and issubclass(result, Exception):
+            with pytest.raises(result):
+                admin.tabulator.set_table("test", "table", "")
+        else:
+            assert admin.tabulator.set_table("test", "table", "") == result
+
+
+@pytest.mark.parametrize(
+    "data,result",
+    [
+        ({"__typename": "BucketConfig"}, None),
+        *MUTATION_ERRORS,
+    ],
+)
+def test_tabulator_rename(data, result):
+    with mock_client(
+        _make_nested_dict("admin.bucket_rename_tabulator_table", data),
+        "bucketTabulatorTableRename",
+        variables={"bucketName": "test", "tableName": "table", "newTableName": "new_table"},
+    ):
+        if isinstance(result, type) and issubclass(result, Exception):
+            with pytest.raises(result):
+                admin.tabulator.rename_table("test", "table", "new_table")
+        else:
+            assert admin.tabulator.rename_table("test", "table", "new_table") == result